+++ crates/universal-sierra-compiler-api/Cargo.toml
@@ -9,3 +9,6 @@ serde.workspace = true
 serde_json.workspace = true
 which.workspace = true
 tempfile.workspace = true
+num-bigint.workspace = true
+cairo-lang-casm.workspace = true
+cairo-lang-sierra.workspace = true
