+++ crates/universal-sierra-compiler-api/src/lib.rs
@@ -1,4 +1,8 @@
 use anyhow::{Context, Result};
+use cairo_lang_casm::hints::Hint;
+use cairo_lang_sierra::program::Program;
+use num_bigint::BigInt;
+use serde::{Deserialize, Serialize};
 use serde_json::Value;
 use std::fmt::Display;
 use std::io::Write;
@@ -9,6 +13,30 @@ use tempfile::Builder;
 pub use command::*;
 
 mod command;
+
+#[derive(Serialize, Deserialize)]
+pub struct AssembledProgramWithDebugInfo {
+    pub assembled_cairo_program: AssembledCairoProgramWithSerde,
+    pub debug_info: Vec<(usize, usize)>,
+}
+
+#[derive(Serialize, Deserialize, Clone)]
+pub struct AssembledCairoProgramWithSerde {
+    pub bytecode: Vec<BigInt>,
+    pub hints: Vec<(usize, Vec<Hint>)>,
+}
+
+pub fn compile_sierra_to_casm(sierra_program: &Program) -> Result<AssembledProgramWithDebugInfo> {
+    let assembled_with_info_raw = compile_sierra(
+        &serde_json::to_value(sierra_program).unwrap(),
+        None,
+        &SierraType::Raw,
+    )?;
+    let assembled_with_info: AssembledProgramWithDebugInfo =
+        serde_json::from_str(&assembled_with_info_raw)?;
+
+    Ok(assembled_with_info)
+}
 
 pub fn compile_sierra(
     sierra_contract_class: &Value,
@@ -52,7 +80,7 @@ pub fn compile_sierra_at_path(
         .output()
         .context(
             "Error while compiling Sierra. \
-            Make sure you have the latest universal-sierra-binary installed. \
+            Make sure you have the latest universal-sierra-compiler binary installed. \
             Contact us if it doesn't help",
         )?;
